// src/types/patient.ts
export type Patient = {
    patientName: string;
    patientUserName: string;
    age: number;
    gender: string;
    tobaccoUsage: string;
    patientGender: string;
    patientSmokingStatus: string;
    systolicBloodPressure: string;
    diastolicBloodPressure: string;
    bloodPressureRecorded: string;
};